requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
responses>=0.24.0
numpy>=1.26.0
pandas>=2.1.0
pytz>=2023.3
